requirements.txt
ftfy==6.2.3
jittor==1.3.8.5
loguru==0.7.2
numpy==1.24.4
pandas==2.2.2
Pillow==10.4.0
PyYAML==6.0.1
regex==2024.4.28
scikit_learn==1.5.1
setuptools==68.2.2
tqdm==4.66.4
